feat: authentication lazy loading
